canned message - view module

+++ app/Http/Controllers/CannedCategoryController.php
@@ -55,9 +55,12 @@ class CannedCategoryController extends Controller
         $count = CannedMessageModel::where('canned_categories_id',$id)->count();
         if($count === 0)
         {
-            $category = CannedCategory::find($id);
-            $category->delete();
-            return response()->json(['success' => true, 'message' => 'Category successfully deleted!']);
+            if($category = CannedCategory::find($id))
+            {
+                $category->delete();
+                return response()->json(['success' => true, 'message' => 'Category successfully deleted!']);
+            }
+            return response()->json(['success' => false, 'message' => 'Token Mismatch','reload' => true]);
         }
         return response()->json(['success' => false, 'message' => 'Category cannot be deleted']);
     }
